+++ frontend/src/pages/Methodology/methodologyComponents/MethodologyCardMenuMobile.tsx
@@ -1,9 +1,4 @@
-import AppBar from '@mui/material/AppBar'
 import Toolbar from '@mui/material/Toolbar'
-import Typography from '@mui/material/Typography'
-import CssBaseline from '@mui/material/CssBaseline'
-import Container from '@mui/material/Container'
-import Box from '@mui/material/Box'
 import { Select, FormControl, MenuItem, InputLabel, Card } from '@mui/material'
 import { links } from './MethodologyCardMenu'
 import { useHistory } from 'react-router-dom'
