+++ frontend/src/pages/Methodology/methodologySections/ChronicDiseaseLink.tsx
@@ -1,26 +1,9 @@
 import styles from '../methodologyComponents/MethodologyPage.module.scss'
-import { urlMap } from '../../../utils/externalUrls'
 import KeyTerms from '../methodologyComponents/KeyTerms'
 import {
   chronicDiseaseDataSources,
   chronicDiseaseDefinitionsArray,
 } from '../methodologyContent/ChronicDiseaseDefinitions'
-import DataTable from '../methodologyComponents/DataTable'
-import {
-  RESOURCES,
-  PDOH_RESOURCES,
-  EQUITY_INDEX_RESOURCES,
-  AIAN_RESOURCES,
-  API_RESOURCES,
-  HISP_RESOURCES,
-  MENTAL_HEALTH_RESOURCES,
-  COVID_RESOURCES,
-  COVID_VACCINATION_RESOURCES,
-  ECONOMIC_EQUITY_RESOURCES,
-  HIV_RESOURCES,
-} from '../../WhatIsHealthEquity/ResourcesData'
-import Resources from '../methodologyComponents/Resources'
-import ConditionVariable from '../methodologyContent/ConditionVariable'
 import { Helmet } from 'react-helmet-async'
 import { Alert, AlertTitle } from '@mui/material'
 import { CodeBlock } from '../methodologyComponents/CodeBlock'
@@ -28,7 +11,7 @@ import {
   percentShareTooltip,
   totalCasesPer100kPeopleTooltip,
 } from '../methodologyContent/TooltipLibrary'
-import AgeAdjustmentExampleTable from '../methodologyComponents/AgeAdjustmentExampleTable'
+import StripedTable from '../methodologyComponents/StripedTable'
 import { DATA_CATALOG_PAGE_LINK } from '../../../utils/internalRoutes'
 import { DATA_SOURCE_PRE_FILTERS } from '../../../utils/urlutils'
 
@@ -41,7 +24,7 @@ const ChronicDiseaseLink = () => {
         </Helmet>
         <h2 className={styles.ScreenreaderTitleHeader}>Chronic Diseases</h2>
         <br />
-        <AgeAdjustmentExampleTable
+        <StripedTable
           id="#categories-table"
           applyThickBorder={false}
           columns={[
@@ -148,7 +131,7 @@ const ChronicDiseaseLink = () => {
         />
 
         <h3 id="#chronic-diseases-data-sources">Data Sources</h3>
-        <AgeAdjustmentExampleTable
+        <StripedTable
           applyThickBorder={false}
           columns={[
             { header: 'Source', accessor: 'source' },
